Pin ICE to the configured port above 32767

+++ src/main/java/dev/waterdog/waterdogpe/network/nethernet/NetherNetInterface.java
@@ -206,8 +206,8 @@ public class NetherNetInterface implements NetworkInterface, SignalingService {
         }
         return config
                 .withEnableIceUdpMux(true)
-                .withPortRangeBegin((short) port)
-                .withPortRangeEnd((short) port);
+                .withPortRangeBegin(port)
+                .withPortRangeEnd(port);
     }
 
     /**
